moved FakeAuthProvider to test-utils
