+++ app/routes.test.js
@@ -1,7 +1,7 @@
 import request from 'supertest';
 import { App } from '../app';
 import { routes } from './routes';
-import { FakeAuthProvider } from './authProvider';
+import { FakeAuthProvider } from './test-utils/FakeAuthProvider';
 import * as homepageContext from './pages/homepage/context';
 import * as viewSolutionController from './pages/view-solution/controller';
 import * as solutionListPageContext from './pages/solutions-list/controller';
